requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==3.0.1
PyMySQL==1.1.0
cryptography==41.0.7
